Warning fixes

+++ src/lib/protocols/kerberos.c
@@ -56,7 +56,7 @@ static int krb_decode_asn1_length(struct ndpi_detection_module_struct *ndpi_stru
 
   *kasn1_offset += value_len;
 
-  return length;
+  return (int)length;
 }
 
 /* Reference: https://en.wikipedia.org/wiki/X.690#Identifier_octets */
@@ -173,7 +173,7 @@ static int krb_decode_asn1_blocks_skip(struct ndpi_detection_module_struct *ndpi
 static void krb_strncpy_lower(char * const dst, size_t dst_siz,
                               char const * const src, size_t src_siz)
 {
-  int i, dst_len = ndpi_min(src_siz, dst_siz - 1);
+  int i, dst_len = (int)ndpi_min(src_siz, dst_siz - 1);
 
    dst[dst_len] = '\0';
